Replace old RNA-seq pipeline with beta
