+++ resolwe_bio/tests/processes/test_expression.py
@@ -363,143 +363,6 @@ class ExpressionProcessorTestCase(KBBioProcessTestCase):
         base = Path("test_featurecounts")
         inputs = base / "inputs"
         outputs = base / "outputs"
-        with self.preparation_stage():
-            annotation_gtf = self.run_process(
-                "upload-gtf",
-                {
-                    "src": str(inputs / "feature_counts hs.gtf.gz"),
-                    "source": "ENSEMBL",
-                    "species": "Homo sapiens",
-                    "build": "GRCh38_ens90",
-                },
-            )
-            annotation_gff3 = self.prepare_annotation_gff()
-
-            bam_single = self.run_process(
-                "upload-bam",
-                {
-                    "src": str(inputs / "reads.bam"),
-                    "species": "Dictyostelium discoideum",
-                    "build": "dd-05-2009",
-                },
-            )
-
-            bam_paired = self.run_process(
-                "upload-bam",
-                {
-                    "src": str(inputs / "feature_counts hs_paired.bam"),
-                    "species": "Homo sapiens",
-                    "build": "GRCh38_ens90",
-                },
-            )
-
-            bam_ucsc = self.run_process(
-                "upload-bam",
-                {
-                    "src": str(inputs / "cuffquant_mapping.bam"),
-                    "species": "Homo sapiens",
-                    "build": "hg19",
-                },
-            )
-
-            annotation_ucsc = self.prepare_annotation(
-                fn=str(inputs / "hg19_chr20_small_modified.gtf.gz"),
-                source="UCSC",
-                species="Homo sapiens",
-                build="hg19",
-            )
-        # test using BAM file containing paired-end reads and a GTF input file
-        expression = self.run_process(
-            "feature_counts",
-            {
-                "aligned_reads": bam_paired.id,
-                "annotation": annotation_gtf.id,
-            },
-        )
-        self.assertFile(
-            expression,
-            "rc",
-            outputs / "feature_counts_out_rc.tab.gz",
-            compression="gzip",
-        )
-        self.assertFile(
-            expression,
-            "cpm",
-            outputs / "feature_counts_out_cpm.tab.gz",
-            compression="gzip",
-        )
-        self.assertFile(
-            expression,
-            "exp",
-            outputs / "feature_counts_out_tpm.tab.gz",
-            compression="gzip",
-        )
-        self.assertFile(
-            expression,
-            "exp_set",
-            outputs / "feature_counts_out_exp_set.txt.gz",
-            compression="gzip",
-        )
-        self.assertJSON(
-            expression,
-            expression.output["exp_set_json"],
-            "",
-            outputs / "feature_counts_exp_set.json.gz",
-        )
-        self.assertFields(expression, "species", "Homo sapiens")
-        self.assertFields(expression, "build", "GRCh38_ens90")
-        self.assertFields(expression, "feature_type", "gene")
-
-        # test using BAM file containing single-end reads and a GFF input file
-        expression = self.run_process(
-            "feature_counts",
-            {
-                "aligned_reads": bam_single.id,
-                "annotation": annotation_gff3.id,
-                "id_attribute": "Parent",
-            },
-        )
-        self.assertFile(
-            expression,
-            "rc",
-            outputs / "feature_counts_out_gff3_rc.tab.gz",
-            compression="gzip",
-        )
-        self.assertFile(
-            expression,
-            "exp",
-            outputs / "feature_counts_out_gff3_tpm.tab.gz",
-            compression="gzip",
-        )
-        self.assertFields(expression, "feature_type", "gene")
-
-        # test using UCSC-derived annotation
-        expression = self.run_process(
-            "feature_counts",
-            {
-                "aligned_reads": bam_ucsc.id,
-                "annotation": annotation_ucsc.id,
-            },
-        )
-        self.assertFile(
-            expression,
-            "rc",
-            outputs / "feature_counts_out_ucsc_rc.tab.gz",
-            compression="gzip",
-        )
-        self.assertFile(
-            expression,
-            "exp",
-            outputs / "feature_counts_out_ucsc_tpm.tab.gz",
-            compression="gzip",
-        )
-
-    @with_resolwe_host
-    @tag_process("feature_counts-beta")
-    def test_feature_counts_beta(self):
-        base = Path("test_featurecounts")
-        inputs = base / "inputs"
-        outputs = base / "outputs"
         input_folder = Path("test_star") / "input"
         with self.preparation_stage():
             annotation_gtf = self.run_process(
@@ -573,7 +436,7 @@ class ExpressionProcessorTestCase(KBBioProcessTestCase):
             )
         # test using BAM file containing paired-end reads and a GTF input file
         expression = self.run_process(
-            "feature_counts-beta",
+            "feature_counts",
             {
                 "aligned_reads": bam_paired.id,
                 "annotation": annotation_gtf.id,
@@ -615,7 +478,7 @@ class ExpressionProcessorTestCase(KBBioProcessTestCase):
 
         # test using BAM file containing single-end reads and a GFF input file
         expression = self.run_process(
-            "feature_counts-beta",
+            "feature_counts",
             {
                 "aligned_reads": bam_single.id,
                 "annotation": annotation_gff3.id,
@@ -637,7 +500,7 @@ class ExpressionProcessorTestCase(KBBioProcessTestCase):
         self.assertFields(expression, "feature_type", "gene")
 
         expression_lanes = self.run_process(
-            "feature_counts-beta",
+            "feature_counts",
             {
                 "aligned_reads": single_lanes.id,
                 "annotation": annotation.id,
@@ -652,7 +515,7 @@ class ExpressionProcessorTestCase(KBBioProcessTestCase):
 
         # test using UCSC-derived annotation
         expression = self.run_process(
-            "feature_counts-beta",
+            "feature_counts",
             {
                 "aligned_reads": bam_ucsc.id,
                 "annotation": annotation_ucsc.id,
@@ -672,7 +535,7 @@ class ExpressionProcessorTestCase(KBBioProcessTestCase):
         )
 
         expression_lanes = self.run_process(
-            "feature_counts-beta",
+            "feature_counts",
             {
                 "aligned_reads": paired_lanes.id,
                 "annotation": annotation.id,
